Code-Update 1: Making  InfluxDb FluxQueries more Flexible and cleaning the Services
Made the FluxQueries more flexible referencing and using the values from the properties.

Clean Coding change making the Service only work as a reading service for the influxDB operations. Every Method should be divided to one goal not many.

Added new enums finalizing all 4 queries. The queries will only work for the weatherstation of FH-Dortmund, but the influxdb can be imported anywhere if properties settings for buckets and measurement are updated. So the Code will be stay flexible even if the influxdb is setup under different buckets.

Improved FluxTableMapper which is now able to map historical data.